+++ src/main/java/com/ykkj/weiyi/quartz/HdTask.java
@@ -55,33 +55,42 @@ public class HdTask
         st = hdService.findProjectInfoMaxInserttime();
         // 配置参数
         Map<String, String> map = new HashMap<String, String>();
+
         map.put("type", "0");
         map.put("startTime", st);
+
+        /*
+         * map.put("startTime", "2015-08-01 00:00:00"); map.put("endTime",
+         * "2016-04-01 00:00:00");
+         */
         String datas = HttpUtils.URLGet(StaticUtils.HD_GETPROJECTINFO, map,
                 "UTF-8");
         JSONObject res = JSONObject.fromObject(datas);
-        JSONArray jsonArray = res.getJSONArray("returnData");
-        Iterator<JSONObject> iterator = jsonArray.iterator();
-        JSONObject jsonObject;
-        ProjectInfoWithBLOBs pwb;
-        while (iterator.hasNext())
+        if ("0".equals(res.getString("status")))
         {
-            jsonObject = iterator.next();
-            pwb = new ProjectInfoWithBLOBs();
-            jsonToObj(pwb, jsonObject);
-            int id = Integer.valueOf(pwb.getId());
-            id++;
-            pwb.setId(String.valueOf(id));
-            if (hdService.findProjectInfoWithBLOBsByID(pwb.getId()) != null)
+            JSONArray jsonArray = res.getJSONArray("returnData");
+            Iterator<JSONObject> iterator = jsonArray.iterator();
+            JSONObject jsonObject;
+            ProjectInfoWithBLOBs pwb;
+            while (iterator.hasNext())
             {
-                // System.out.println("id:" + pwb.getId() + "重复，修改中....");
-                hdService.deleteProjectInfoByID(pwb.getId());
-                // System.out.println("修改成功");
+                jsonObject = iterator.next();
+                pwb = new ProjectInfoWithBLOBs();
+                jsonToObj(pwb, jsonObject);
+                if (hdService.findProjectInfoWithBLOBsByID(pwb.getId()) != null)
+                {
+                    // System.out.println("id:" + pwb.getId() + "重复，修改中....");
+                    hdService.deleteProjectInfoByID(pwb.getId());
+                    // System.out.println("修改成功");
+                }
+                pwb.setCreatetime(new Timestamp(new Date().getTime()));
+                hdService.insert(pwb);
             }
-            pwb.setCreatetime(new Timestamp(new Date().getTime()));
-            hdService.insert(pwb);
         }
-
+        else
+        {
+            System.out.println(res);
+        }
         logger.info("getProjectInfo 结束时间:" + sf.format(new Date()));
     }
 
@@ -101,29 +110,33 @@ public class HdTask
         String datas = HttpUtils.URLGet(StaticUtils.HD_GETDIRECTINFO, map,
                 "UTF-8");
         JSONObject res = JSONObject.fromObject(datas);
-        JSONArray jsonArray = res.getJSONArray("returnData");
-        Iterator<JSONObject> iterator = jsonArray.iterator();
-        JSONObject jsonObject;
-        DirectInfoWithBLOBs dwb;
-        while (iterator.hasNext())
+        if ("0".equals(res.getString("status")))
         {
-            jsonObject = iterator.next();
-            dwb = new DirectInfoWithBLOBs();
-            jsonToObj(dwb, jsonObject);
-            int id = Integer.valueOf(dwb.getId());
-            id++;
-            dwb.setId(String.valueOf(id));
-            if (hdService.findDirectInfoByID(dwb.getId()) != null)
+            JSONArray jsonArray = res.getJSONArray("returnData");
+            Iterator<JSONObject> iterator = jsonArray.iterator();
+            JSONObject jsonObject;
+            DirectInfoWithBLOBs dwb;
+            while (iterator.hasNext())
             {
-                /*
-                 * System.out.println("DirectInfo id:" + dwb.getId() +
-                 * "重复，修改中....");
-                 */
-                hdService.deleteDirectInfoByID(dwb.getId());
-                /* System.out.println("修改成功"); */
+                jsonObject = iterator.next();
+                dwb = new DirectInfoWithBLOBs();
+                jsonToObj(dwb, jsonObject);
+                if (hdService.findDirectInfoByID(dwb.getId()) != null)
+                {
+                    /*
+                     * System.out.println("DirectInfo id:" + dwb.getId() +
+                     * "重复，修改中....");
+                     */
+                    hdService.deleteDirectInfoByID(dwb.getId());
+                    /* System.out.println("修改成功"); */
+                }
+                dwb.setCreatetime(new Timestamp(new Date().getTime()));
+                hdService.insertDirectInfo(dwb);
             }
-            dwb.setCreatetime(new Timestamp(new Date().getTime()));
-            hdService.insertDirectInfo(dwb);
+        }
+        else
+        {
+            System.out.println(res);
         }
         logger.info("getDirectInfo 结束时间:" + sf.format(new Date()));
     }
@@ -144,29 +157,33 @@ public class HdTask
         String datas = HttpUtils.URLGet(StaticUtils.HD_GETSUPERVISIONINFO, map,
                 "UTF-8");
         JSONObject res = JSONObject.fromObject(datas);
-        JSONArray jsonArray = res.getJSONArray("returnData");
-        Iterator<JSONObject> iterator = jsonArray.iterator();
-        JSONObject jsonObject;
-        SupervisionInfoWithBLOBs swb;
-        while (iterator.hasNext())
+        if ("0".equals(res.getString("status")))
         {
-            jsonObject = iterator.next();
-            swb = new SupervisionInfoWithBLOBs();
-            jsonToObj(swb, jsonObject);
-            int id = Integer.valueOf(swb.getId());
-            id++;
-            swb.setId(String.valueOf(id));
-            if (hdService.findSupervisionInfoByID(swb.getId()) != null)
+            JSONArray jsonArray = res.getJSONArray("returnData");
+            Iterator<JSONObject> iterator = jsonArray.iterator();
+            JSONObject jsonObject;
+            SupervisionInfoWithBLOBs swb;
+            while (iterator.hasNext())
             {
-                /*
-                 * System.out.println("getSupervisionInfo id:" + swb.getId() +
-                 * "重复，修改中....");
-                 */
-                hdService.deleteSupervisionInfoByID(swb.getId());
-                /* System.out.println("修改成功"); */
+                jsonObject = iterator.next();
+                swb = new SupervisionInfoWithBLOBs();
+                jsonToObj(swb, jsonObject);
+                if (hdService.findSupervisionInfoByID(swb.getId()) != null)
+                {
+                    /*
+                     * System.out.println("getSupervisionInfo id:" + swb.getId()
+                     * + "重复，修改中....");
+                     */
+                    hdService.deleteSupervisionInfoByID(swb.getId());
+                    /* System.out.println("修改成功"); */
+                }
+                swb.setCreatetime(new Timestamp(new Date().getTime()));
+                hdService.insertSupervisionInfo(swb);
             }
-            swb.setCreatetime(new Timestamp(new Date().getTime()));
-            hdService.insertSupervisionInfo(swb);
+        }
+        else
+        {
+            System.out.println(res);
         }
         logger.info("getSupervisionInfo 结束时间:" + sf.format(new Date()));
     }
@@ -187,29 +204,34 @@ public class HdTask
         String datas = HttpUtils.URLGet(StaticUtils.HD_GETSECTINFO, map,
                 "UTF-8");
         JSONObject res = JSONObject.fromObject(datas);
-        JSONArray jsonArray = res.getJSONArray("returnData");
-        Iterator<JSONObject> iterator = jsonArray.iterator();
-        JSONObject jsonObject;
-        SectInfoWithBLOBs swb;
-        while (iterator.hasNext())
+        if ("0".equals(res.getString("status")))
         {
-            jsonObject = iterator.next();
-            swb = new SectInfoWithBLOBs();
-            jsonToObj(swb, jsonObject);
-            int id = Integer.valueOf(swb.getId());
-            id++;
-            swb.setId(String.valueOf(id));
-            if (hdService.findSectInfoByID(swb.getId()) != null)
+            JSONArray jsonArray = res.getJSONArray("returnData");
+            Iterator<JSONObject> iterator = jsonArray.iterator();
+            JSONObject jsonObject;
+            SectInfoWithBLOBs swb;
+            while (iterator.hasNext())
             {
-                /*
-                 * System.out.println("DirectInfo id:" + swb.getId() +
-                 * "重复，修改中....");
-                 */
-                hdService.deleteSectInfoByID(swb.getId());
-                /* System.out.println("修改成功"); */
+                jsonObject = iterator.next();
+                swb = new SectInfoWithBLOBs();
+                jsonToObj(swb, jsonObject);
+
+                if (hdService.findSectInfoByID(swb.getId()) != null)
+                {
+                    /*
+                     * System.out.println("DirectInfo id:" + swb.getId() +
+                     * "重复，修改中....");
+                     */
+                    hdService.deleteSectInfoByID(swb.getId());
+                    /* System.out.println("修改成功"); */
+                }
+                swb.setCreatetime(new Timestamp(new Date().getTime()));
+                hdService.insertSectInfo(swb);
             }
-            swb.setCreatetime(new Timestamp(new Date().getTime()));
-            hdService.insertSectInfo(swb);
+        }
+        else
+        {
+            System.out.println(res);
         }
         logger.info("getSectInfo 结束时间:" + sf.format(new Date()));
     }
@@ -249,7 +271,8 @@ public class HdTask
                     field.set(obj, jsonObject.getString(name));
                     // field.set(obj, reqVal);
                 }
-                else if (fieldType == Integer.class && jsonObject.has(name))
+                else if (fieldType == Integer.class && jsonObject.has(name)
+                        && !"".equals(jsonObject.getString(name)))
                 {
                     field.set(obj, jsonObject.getInt(name));
                     // field.set(obj, NumberUtils.toInt(reqVal));
@@ -264,8 +287,10 @@ public class HdTask
                     dtemp = jsonObject.getDouble(name);
                     field.set(obj, Float.valueOf(jsonObject.getString(name)));
                 }
-                else if (fieldType == BigDecimal.class && jsonObject.has(name))
+                else if (fieldType == BigDecimal.class && jsonObject.has(name)
+                        && !"".equals(jsonObject.getString(name)))
                 {
+
                     field.set(obj, new BigDecimal(jsonObject.getString(name)));
                 }
                 else if (fieldType == Boolean.class && jsonObject.has(name))
@@ -273,9 +298,19 @@ public class HdTask
                     field.set(obj, jsonObject.getBoolean(name));
                     // field.set(obj, Boolean.valueOf(reqVal));
                 }
-                else if (fieldType == Date.class && jsonObject.has(name))
+                else if (fieldType == Date.class && jsonObject.has(name)
+                        && !"".equals(jsonObject.getString(name)))
                 {// 时间
-                    field.set(obj, sdf.parse(jsonObject.getString(name)));
+                    if (jsonObject.getString(name).length() == 10)
+                    {
+                        SimpleDateFormat sf1 = new SimpleDateFormat(
+                                "yyyy-MM-dd");
+                        field.set(obj, sf1.parse(jsonObject.getString(name)));
+                    }
+                    else
+                    {
+                        field.set(obj, sdf.parse(jsonObject.getString(name)));
+                    }
                 }
             }
         }
